Override SvelteHtmlAttribute.processDeclarations
